requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary
python-dotenv
PyJWT==2.8.0
bcrypt==4.1.2
Flask-Mail==0.9.1
stripe==7.9.0
mercadopago==2.2.1
